requirements.txt
streamlit==1.32.0
google-generativeai>=0.8.3
fpdf==1.7.2
python-dotenv==1.0.0